fix: sort chunksArray before merge

+++ app.js
@@ -6,6 +6,7 @@ const path = require('path');
 const fs = require('fs-extra');
 const crypto = require('crypto');
 const koaBody = require('koa-body');
+const chalk = require('chalk');
 const uploadPath = path.join(__dirname, 'upload');
 const uploadTempPath = path.join(uploadPath, 'temp');
 const upload = multer({ dest: uploadTempPath });
@@ -33,7 +34,7 @@ router.post('/file/upload', upload.single('file'), async (ctx, next) => {
   const buffer = fs.readFileSync(ctx.req.file.path);
   const hash = crypto.createHash('md5').update(buffer).digest('hex');
   if (hash !== chunkHash) {
-    console.log('分片校验失败');
+    console.log(chalk('分片校验失败'));
     ctx.status = 500;
     ctx.res.end('分片校验失败');
   }
@@ -55,10 +56,14 @@ router.post('/file/merge_chunks', async (ctx, next) => {
   const filePath = path.join(uploadPath, name);
   // 读取所有的chunks 文件名存放在数组中
   const chunks = fs.readdirSync(chunksPath);
+  chunks.sort((a, b) => {
+    const reg = /(?<=-)\d+$/g;
+    return a.match(reg)[0] - b.match(reg)[0];
+  });
   // 创建存储文件
   fs.writeFileSync(filePath, '');
   if (chunks.length !== total || chunks.length === 0) {
-    ctx.status = 200;
+    ctx.status = 500;
     ctx.res.end('切片文件数量不符合');
     return;
   }
@@ -68,9 +73,20 @@ router.post('/file/merge_chunks', async (ctx, next) => {
     // 删除本次使用的chunk
     fs.unlinkSync(path.join(chunksPath, chunk));
   });
+  const hash = crypto
+    .createHash('md5')
+    .update(fs.readFileSync(filePath))
+    .digest('hex');
+  if (hash !== fileHash) {
+    ctx.status = 200;
+    ctx.type = 'text/plain; charset=utf-8';
+    ctx.res.end('文件完整性校验失败');
+    return;
+  }
   fs.rmdirSync(chunksPath);
   // 文件合并成功
   ctx.status = 200;
+  ctx.type = 'text/plain; charset=utf-8';
   ctx.res.end('合并成功');
 });
 
@@ -78,5 +94,6 @@ app.use(router.routes());
 app.use(router.allowedMethods());
 app.use(serve(__dirname + '/public')); // 静态文件服务
 app.listen(9000, () => {
-  console.log('listening port:9000');
+  // console.log('listening localhost:9000');
+  console.log(`  - Local:   ${chalk.cyan('http://localhost:9000')}`);
 });
